refactor test_toynet into test_toynet & test_submodules
