+++ tests/test_toynet.py
@@ -25,63 +25,13 @@ def test_hello_reclass(client):
     rv = client.get('/')
     assert b'Hello, Reclass!' in rv.data
 
-def test_valueById_get(client):
-    """Check the app is running"""
+def test_user_post(client):
+    """Check users can be added"""
 
-    rv = client.get('/api/value/5004')
-    assert rv.status_code == 200
-    rv_json = json.loads(rv.data.decode('utf-8'))
+    assert True
 
-    assert rv_json['value'] == 'Loyalty'
-    assert len(rv_json['inspiration']) == 1
+def test_userByUsername_get(client):
+    """Check users can be retrieved by username"""
 
-    inspiration = rv_json['inspiration'][0]
-    assert inspiration['organization'] == 'U.S. Army'
-    assert inspiration['definition'][:21] == 'Bear true faith and a'
-    assert inspiration['definition'][-21:] == 'loyalty to your unit.'
+    assert True
 
-    rv = client.get('/api/value/1')
-    assert rv.status_code == 404
-    rv_json = json.loads(rv.data.decode('utf-8'))
-    assert rv_json['message'] == 'value ID 1 does not exist'
-
-def test_quiz_get(client):
-    """Check the app is running"""
-
-    rv = client.get('/api/quiz/4002')
-    assert rv.status_code == 200
-    rv_json = json.loads(rv.data.decode('utf-8'))
-
-    assert rv_json[0]["question"] == 'Which Protocol do Internet Browsers (Chrome, Firefox, Internet Explorer) use?'
-    assert len(rv_json) == 10
-    assert len(rv_json[0]['options']) == 4
-
-    assert rv_json[1]["options"][0] == "Physical"
-    assert rv_json[1]["answer"] == 1
-    assert rv_json[2]["options"][3] == "IPP"
-    assert rv_json[2]["answer"] == 2
-
-    rv = client.get('/api/quiz/2')
-    assert rv.status_code == 404
-    rv_json = json.loads(rv.data.decode('utf-8'))
-    assert rv_json['message'] == 'quiz ID 2 doesn\'t exist'
-
-def test_survey_get(client):
-    """Check the app is running"""
-
-    rv = client.get('/api/survey/6001')
-    assert rv.status_code == 200
-    rv_json = json.loads(rv.data.decode('utf-8'))
-
-    assert rv_json[0]["question"] == 'What is your first name?'
-    assert len(rv_json) == 7
-    assert len(rv_json[1]['options']) == 4
-
-    assert rv_json[1]["options"][0] == "As of now, I don't plan on it."
-    assert rv_json[3]["options"][4] == "Professional Experience"
-    assert rv_json[6]["unit"] == "months"
-
-    rv = client.get('/api/survey/4')
-    assert rv.status_code == 404
-    rv_json = json.loads(rv.data.decode('utf-8'))
-    assert rv_json['message'] == 'survey ID 4 doesn\'t exist'
